requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
chardet==3.0.4
discord.py==0.16.8
multidict==4.3.1
olefile==0.45.1
Pillow==4.1.1
redis==2.10.6
websockets==3.4
